requirements.txt
flask
cased-kit
python-dotenv
openai
